backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.9.1",
    "authlib==1.6.1",
    "cffi==1.17.1",
    "cryptography==45.0.5",
    "django==5.2.4",
    "django-cors-headers>=4.7.0",
    "django-guardian>=3.0.3",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "djangorestframework==3.16.0",
    "djangorestframework-guardian>=0.4.0",
    "gunicorn==23.0.0",
    "packaging==25.0",
    "pillow>=11.3.0",
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "psycopg-pool==3.2.6",
    "pycparser==2.22",
    "pyjwt==2.10.1",
    "python-dotenv==1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.4",
    "sqlparse==0.5.3",
    "typing-extensions==4.14.1",
]
